LayoutTests/http/tests/webgpu/webgpu/api/validation/create_pipeline.html


                <script>
                    if (window.testRunner)
                        testRunner.waitUntilDone();
                    window.location.href = "/webgpu/cts.https.html?q=webgpu:api,validation,create_pipeline:*";
                </script>
            
